Eliminar a merda toda e por a merda nova
Esper que tenha funcionado :)

+++ app/Http/Controllers/FilmesController.php
@@ -7,18 +7,18 @@ use App\Models\Genero;
 use Illuminate\Http\Request;
 use Illuminate\View\View;
 use Illuminate\Support\Facades\Log;
+use App\Http\Requests\FilmeRequest;
+use Illuminate\Support\Facades\DB;
 
 class FilmesController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index() : View
+    public function index(): View
     {
-        $filmes = Filme::all();
-        debug($filmes);
-        Log::debug('Cursos has been loaded on the controller.', ['$allCursos' => $filmes]);
-        return view('filmes.index')->with('filmes', $filmes);
+        $filmes = Filme::paginate(10);
+        return view('filmes.index', compact('filmes'));
     }
 
     /**
@@ -26,8 +26,9 @@ class FilmesController extends Controller
      */
     public function create()
     {
-        $filme = new Filme();
         $generos = Genero::all();
+
+        $filme = new Filme();
 
         return view('filmes.create', compact('filme', 'generos'));
     }
@@ -35,20 +36,14 @@ class FilmesController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(Request $request)
+    public function store(FilmeRequest $request)
     {
-        $validated = $request->validate([
-            'titulo' => 'required',
-            'genero_code' => 'required|exists:generos,code',
-            'ano' => 'required',
-            'cartaz_url' => 'required',
-            'sumario' => 'required',
-            'trailer_url' => 'required'
-            ]);
-
-        Filme::create($validated);
-        return redirect()->route('filmes.index');
-
+        $newFilme = Filme::create($request->validated());
+        $url = route('filmes.show', ['filme' => $newFilme]);
+        $htmlMessage = "Filme <a href='$url'>#{$newFilme->id}</a><strong>\"{$newFilme->titulo}\"</strong> foi criada com sucesso!";
+        return redirect()->route('filmes.index')
+            ->with('alert-msg', $htmlMessage)
+            ->with('alert-type', 'success');
     }
 
     /**
@@ -56,7 +51,8 @@ class FilmesController extends Controller
      */
     public function show(Filme $filme)
     {
-        return view('filmes.show', compact('filme'));
+        $generos = Genero::all();
+        return view('filmes.show', compact('filme', 'generos'));
     }
 
     /**
@@ -71,10 +67,14 @@ class FilmesController extends Controller
     /**
      * Update the specified resource in storage.
      */
-    public function update(Request $request, Filme $filme)
+    public function update(FilmeRequest $request, Filme $filme)
     {
-        $filme->update($request->all());
-        return redirect()->route('filmes.index');
+        $filme->update($request->validated());
+        $url = route('filmes.show', ['filme' => $filme]);
+        $htmlMessage = "filme <a href='$url'>#{$filme->id}</a><strong>\"{$filme->titulo}\"</strong> foi alterada com sucesso!";
+        return redirect()->route('filmes.index')
+            ->with('alert-msg', $htmlMessage)
+            ->with('alert-type', 'success');
     }
 
     /**
@@ -82,7 +82,37 @@ class FilmesController extends Controller
      */
     public function destroy(Filme $filme)
     {
-        $filme->delete();
-        return redirect()->route('filmes.index');
+        try {
+            $totalSessoes = DB::scalar('select count(*)
+            from sessoes where filme_id = ?', [$filme->id]);
+            if ($totalSessoes == 0) {
+                $filme->delete();
+                $alertType = 'success';
+                $htmlMessage = "Filme #{$filme->id}
+            <strong>\"{$filme->titulo}\"</strong> foi apagado com sucesso!";
+            } else {
+                $url = route('filmes.show', ['filme' => $filme]);
+                $alertType = 'warning';
+                $sessoesSTR = $totalSessoes > 0 ?
+                    ($totalSessoes == 1 ?
+                        "1 sessão relacionada ao filme" :
+                        "$totalSessoes sessões relacionadas ao filme") :
+                    "";
+
+                $htmlMessage = "Filme <a href='$url'>#{$filme->id}</a>
+            <strong>\"{$filme->titulo}\"</strong>
+            não pode ser apagada porque há $sessoesSTR!
+            ";
+            }
+        } catch (\Exception $error) {
+            $url = route('filmes.show', ['filme' => $filme]);
+            $htmlMessage = "Não foi possível apagar o filme
+            <a href='$url'>#{$filme->id}</a>
+            <strong>\"{$filme->titulo}\"</strong> porque ocorreu um erro!";
+            $alertType = 'danger';
+        }
+        return redirect()->route('filmes.index')
+            ->with('alert-msg', $htmlMessage)
+            ->with('alert-type', $alertType);
     }
 }
